Send PFS update after withdraw
