+++ raiden/raiden_service.py
@@ -36,9 +36,7 @@ from raiden.exceptions import (
 from raiden.messages import (
     LockedTransfer,
     Message,
-    RequestMonitoring,
     SignedMessage,
-    UpdatePFS,
     lockedtransfersigned_from_message,
     message_from_sendevent,
 )
@@ -48,12 +46,13 @@ from raiden.network.proxies.service_registry import ServiceRegistry
 from raiden.network.proxies.token_network_registry import TokenNetworkRegistry
 from raiden.network.proxies.user_deposit import UserDeposit
 from raiden.raiden_event_handler import EventHandler
-from raiden.settings import MEDIATION_FEE, MONITORING_MIN_CAPACITY, MONITORING_REWARD
+from raiden.services import update_services_from_balance_proof
+from raiden.settings import MEDIATION_FEE
 from raiden.storage import sqlite, wal
 from raiden.storage.serialization import JSONSerializer
 from raiden.storage.wal import WriteAheadLog
 from raiden.tasks import AlarmTask
-from raiden.transfer import channel, node, views
+from raiden.transfer import node, views
 from raiden.transfer.architecture import Event as RaidenEvent, StateChange
 from raiden.transfer.identifiers import CanonicalIdentifier
 from raiden.transfer.mediated_transfer.events import SendLockedTransfer
@@ -64,13 +63,7 @@ from raiden.transfer.mediated_transfer.state_change import (
     ActionInitTarget,
 )
 from raiden.transfer.mediated_transfer.tasks import InitiatorTask
-from raiden.transfer.state import (
-    BalanceProofSignedState,
-    BalanceProofUnsignedState,
-    ChainState,
-    HopState,
-    PaymentNetworkState,
-)
+from raiden.transfer.state import ChainState, HopState, PaymentNetworkState
 from raiden.transfer.state_change import (
     ActionChangeNodeNetworkState,
     ActionChannelWithdraw,
@@ -78,7 +71,7 @@ from raiden.transfer.state_change import (
     Block,
     ContractReceiveNewPaymentNetwork,
 )
-from raiden.utils import create_default_identifier, lpex, pex, random_secret, to_rdn
+from raiden.utils import create_default_identifier, lpex, pex, random_secret
 from raiden.utils.runnable import Runnable
 from raiden.utils.signer import LocalSigner, Signer
 from raiden.utils.typing import (
@@ -224,104 +217,6 @@ class PaymentStatus(NamedTuple):
 
     def matches(self, token_network_address: TokenNetworkAddress, amount: PaymentAmount) -> bool:
         return token_network_address == self.token_network_address and amount == self.amount
-
-
-def update_services_from_balance_proof(
-    raiden: "RaidenService",
-    chain_state: "ChainState",
-    balance_proof: Union[BalanceProofSignedState, BalanceProofUnsignedState],
-) -> None:
-    update_path_finding_service_from_balance_proof(
-        raiden=raiden, chain_state=chain_state, new_balance_proof=balance_proof
-    )
-    if isinstance(balance_proof, BalanceProofSignedState):
-        update_monitoring_service_from_balance_proof(
-            raiden=raiden,
-            chain_state=chain_state,
-            new_balance_proof=balance_proof,
-            monitoring_service_contract_address=raiden.default_msc_address,
-        )
-
-
-def update_path_finding_service_from_balance_proof(
-    raiden: "RaidenService",
-    chain_state: "ChainState",
-    new_balance_proof: Union[BalanceProofSignedState, BalanceProofUnsignedState],
-) -> None:
-    channel_state = views.get_channelstate_by_canonical_identifier(
-        chain_state=chain_state, canonical_identifier=new_balance_proof.canonical_identifier
-    )
-    network_address = new_balance_proof.canonical_identifier.token_network_address
-    error_msg = (
-        f"tried to send a balance proof in non-existant channel "
-        f"token_network_address: {pex(network_address)} "
-    )
-    assert channel_state is not None, error_msg
-    if isinstance(new_balance_proof, BalanceProofSignedState):
-        assert channel_state.partner_state.balance_proof == new_balance_proof
-    else:  # BalanceProofUnsignedState
-        assert channel_state.our_state.balance_proof == new_balance_proof
-
-    msg = UpdatePFS.from_channel_state(channel_state)
-    msg.sign(raiden.signer)
-    raiden.transport.send_global(constants.PATH_FINDING_BROADCASTING_ROOM, msg)
-    log.debug("Sent a PFS Update", message=msg, balance_proof=new_balance_proof)
-
-
-def update_monitoring_service_from_balance_proof(
-    raiden: "RaidenService",
-    chain_state: "ChainState",
-    new_balance_proof: BalanceProofSignedState,
-    monitoring_service_contract_address: Address,
-) -> None:
-    if raiden.config["services"]["monitoring_enabled"] is False:
-        return
-
-    channel_state = views.get_channelstate_by_canonical_identifier(
-        chain_state=chain_state, canonical_identifier=new_balance_proof.canonical_identifier
-    )
-
-    msg = (
-        f"Failed to update monitoring service due to inability to find "
-        f"channel: {new_balance_proof.channel_identifier} "
-        f"token_network_address: {pex(new_balance_proof.token_network_address)}."
-    )
-    assert channel_state, msg
-
-    balance = channel.get_balance(
-        sender=channel_state.our_state, receiver=channel_state.partner_state
-    )
-
-    if balance < MONITORING_MIN_CAPACITY:
-        log.warn(
-            f"Skipping update to Monitoring service. "
-            f"Available balance of {balance} is less than configured "
-            f"minimum capacity of {MONITORING_MIN_CAPACITY}"
-        )
-        return
-
-    assert raiden.user_deposit is not None
-    rei_balance = raiden.user_deposit.effective_balance(raiden.address, "latest")
-    if rei_balance < MONITORING_REWARD:
-        rdn_balance = to_rdn(rei_balance)
-        rdn_reward = to_rdn(MONITORING_REWARD)
-        log.warn(
-            f"Skipping update to Monitoring service. "
-            f"Your deposit balance {rdn_balance} is less than "
-            f"the required monitoring service reward of {rdn_reward}"
-        )
-        return
-
-    log.info(
-        "Received new balance proof, creating message for Monitoring Service.",
-        balance_proof=new_balance_proof,
-    )
-
-    monitoring_message = RequestMonitoring.from_balance_proof_signed_state(
-        new_balance_proof, MONITORING_REWARD, monitoring_service_contract_address
-    )
-    monitoring_message.sign(raiden.signer)
-    raiden.transport.send_global(constants.MONITORING_BROADCASTING_ROOM, monitoring_message)
 
 
 class RaidenService(Runnable):
